requirements.txt
Django==1.10.6
factory-boy==2.8.1